Seção 5: Projeto: Backend com GraphQL - Aulas 81 a 86

+++ projeto-final/backend/resolvers/Mutation/usuario.js
@@ -20,7 +20,9 @@ const mutations = {
     });
   },
 
-  async novoUsuario(_, { dados }) {
+  async novoUsuario(_, { dados }, ctx) {
+    ctx && ctx.validarAdmin();
+
     try {
       const idsPerfis = [];
 
@@ -57,7 +59,9 @@ const mutations = {
       throw new Error(e.sqlMessage);
     }
   },
-  async excluirUsuario(_, args) {
+  async excluirUsuario(_, args, ctx) {
+    ctx && ctx.validarAdmin();
+
     try {
       const usuario = await obterUsuario(_, args);
       if (usuario) {
@@ -74,12 +78,14 @@ const mutations = {
       throw new Error(e.sqlMessage);
     }
   },
-  async alterarUsuario(_, { filtro, dados }) {
+  async alterarUsuario(_, { filtro, dados }, ctx) {
+    ctx && ctx.validarUsuarioFiltro(filtro);
+
     try {
       const usuario = await obterUsuario(_, { filtro });
       if (usuario) {
         const { id } = usuario;
-        if (dados.perfis) {
+        if (ctx.admin && dados.perfis) {
           await db('usuarios_perfis')
             .where({ usuario_id: id })
             .delete();
